fix(苹果构建): 修复浮点整数转换的符号类型错误
修改 guest/aarch64/scalar-fp-conversion.h，将位移差值中的尾数位宽显式转换为 int，与指数比较保持一致；位宽仅为 23 或 52，保持已有转换语义。

修复 Xcode Cloud Build 446 在 Apple core gate 的 -Werror=-Wsign-conversion 下停止构建。原生 Meson 测试 235 项全部通过，iPhoneOS arm64 严格核心编译通过。

+++ guest/aarch64/scalar-fp-conversion.h
@@ -98,8 +98,8 @@ aarch64_scalar_fp_to_integer(qword_t source, byte_t source_width,
         } else {
             qword_t significand = (UINT64_C(1) << format.fraction_bits) | number.fraction;
             qword_t magnitude = exponent >= (int) format.fraction_bits ?
-                    significand << (exponent - format.fraction_bits) :
-                    significand >> (format.fraction_bits - exponent);
+                    significand << (exponent - (int) format.fraction_bits) :
+                    significand >> ((int) format.fraction_bits - exponent);
             invalid = magnitude > limit;
             value = invalid ? limit : number.sign ? 0 - magnitude : magnitude;
         }
